Update default.rb
Added a condition to handle default attribute for the sftp subsystem that maps to the incorrect location on Ubuntu 14.04. I added this change so the end user does not need to be aware when un-commenting the sftp subsystem attribute in attribute/default.rb will not correctly enable sftp in the sshd_config file.