Fix missing /var/run/sshd directory in docker containers
When running the openssh cookbook in an environment (i.e. docker containers or kitchen-docker) that doesn't use the init system the first run always fails. This is because it is missing the /var/run/sshd priviledged directory. These changess will ensure that the /var/run/sshd directory is present before testing the configuration. Whenever openssh is ran from init it works fine the first time.

Signed-off-by: Kevin Harriss <special.kevin@gmail.com>